Start merging integrator_type (#1239)
We start with the functions that don't conflict, and will follow up in future PRs with changes that harmonize the analogous functions and move them into integrator_type.H.
