+++ integration/Make.package
@@ -14,6 +14,7 @@ endif
 
 CEXE_headers += integrator.H
 CEXE_headers += integrator_data.H
+CEXE_headers += integrator_type.H
 
 ifeq ($(USE_SIMPLIFIED_SDC), TRUE)
   CEXE_headers += integrator_rhs_simplified_sdc.H
