+++ integration/QSS/actual_integrator.H
@@ -46,6 +46,12 @@ void initialize_state (BurnT& state)
     state.time = 0.0;
     state.n_rhs = 0;
     state.n_step = 0;
+
+    // Initialize ydot to zero for Strang burn.
+
+    for (int n = 0; n < SVAR; ++n) {
+        state.ydot_a[n] = 0;
+    }
 }
 
 template <typename BurnT>
